add some stuff

+++ src/main/java/za/co/priscadivineapp/ecom/model/Item.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Item {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private Long id;
     @Column(nullable = false, unique = true)
     private String name;
     private Double price;
@@ -55,11 +55,11 @@ public class Item {
         this.createdAt = createdAt;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
